hooks: move hooks to client
